requirements.txt
mastodon.py
openai
newspaper3k
